fix bug

+++ crawler.py
@@ -27,7 +27,7 @@ user_agents = ['Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KH
                'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.6; rv,2.0.1) Gecko/20100101 Firefox/4.0.1',\
                'Mozilla/5.0 (Windows NT 6.1; rv,2.0.1) Gecko/20100101 Firefox/4.0.1',\
                'Mozilla/5.0 (Windows NT 6.1; rv,2.0.1) Gecko/20100101 Firefox/4.0.1']
-ss_proxies = {"http": "127.0.0.1:8118", "https": "127.0.0.1:8118"}
+ss_proxies = {"http": "127.0.0.1:1080", "https": "127.0.0.1:1080"}
 proxies = {
     'http': 'socks5://user:pass@104.194.72.216:443',
     'https': 'socks5://user:pass@host:port'
@@ -220,14 +220,15 @@ def store_d_html_text_single(data, prefix='./webpage/'):
     service_args.append('--load-images=no')  ##关闭图片加载
     service_args.append('--disk-cache=yes')  ##开启缓存
     driver = webdriver.PhantomJS(executable_path=phantomjs_path, service_args=service_args)
-    driver.set_page_load_timeout(30)
+    driver.set_page_load_timeout(240)
     for pid, url in data:
         try:
-            driver.get(url)
             filename = hashlib.md5(url.encode('utf-8')).hexdigest()
-            with codecs.open(prefix + filename, 'w', 'utf-8') as f:
-                f.write(driver.page_source)
-            driver.get("about:blank")
+            if not os.path.isfile(prefix + filename):
+                driver.get(url)
+                with codecs.open(prefix + filename, 'w', 'utf-8') as f:
+                    f.write(driver.page_source)
+                driver.get("about:blank")
         except Exception as e:  
             pass
 
@@ -240,7 +241,8 @@ def store_d_html_text_multi(data, prefix='./webpage/', threads_num=10):
     threads = []
     id_url = [[] for i in range(threads_num)]
     for i, r in data.iterrows():
-        id_url[i % threads_num].append((r['id'], r['homepage']))
+        if i < 5000:
+            id_url[i % threads_num].append((r['id'], r['homepage']))
     for i in id_url:
         t = threading.Thread(target=store_d_html_text_single,
         args=(i, prefix))
